Fix `flatReaddir` ignoring include condition on symbolic files
This resolves an issue introduced in #52361 where `shouldOmit` is overwritten based on if it is a directory or not without considering include conditions.

In an environment where the the application being built is orchestrated and managed by bazel, all files are symbolic links and thus without this the file includes predicate checks aren't considered - and the build fails under certain conditions.