Ignore postponed rejections
We also don't warn for awaiting a Promise late. That's a fine pattern in
Next.js because it allows for fetching early and awaiting late.

If a Promise does reject before that we still log an error which is fine
because an error actually did happen even if we didn't await it.

We ignore postpones though because they are not unexpected.